Rename to VotingReasonsQueryOptions
